Function to display word place holder into the HTML.
displayState function was added to game.js. The function loops through the randomly selected US State and displays a underscore for each letter in the state. inside the loop is a if statement looking for spaces to place a space in the html instead of a underscore.
After the if statement a span is crated with two attributes id and data-letter, both set to the letter currently in the loop. Then the element is appeneded to the element with class selectedState