Added 1 function, 3 event listeners
onGuessInput function compares user key input and checks if input is a letter, if not then alert user to press a input a letter. Then checks if input is in the US State and Not a repeat incorrect guess, if input is not in the State and is not a repeat incorrect input then it will display incorrect guess on screen, add to incorrect guess array, and reduce total attemps by 1. The function still needs a conditional to check for correct letter input and display the letter in the letter holders.

Added and eventlistener to listen for keypress and pass the keypress to the onGuessInput function then update the total attempt number.